fix: remove duplicate SearchLink component and consolidate to path-based navigation
