+++ src/components/Loader/SearchLink.tsx
@@ -26,7 +26,7 @@ export const SearchLink: React.FC<Props> = ({
   };
 
   const searchString = getSearchWith(searchParams, params);
-  const to = searchString ? `/people?${searchString}` : '/people';
+  const to = searchString ? `/people?${searchString}` : '/people'; // ✅ Path-based URL
 
   return (
     <Link to={to} className={className} onClick={handleClick} data-cy={dataCy}>
